moving testing insfrastructure to lake so that test functions can be installed and used in different files and arnet
